Rename folder, ignore linting on new files
