+++ packages/commerce-sdk-react/.eslintignore
@@ -5,3 +5,4 @@ generator-assets
 vendor
 temp
 temp_static
+src/components/ShopperContext
