+++ packages/commerce-sdk-react/src/components/index.ts
@@ -5,4 +5,4 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 export * from './ShopperExperience'
-export * from './ShopperContext'
+export * from './StorefrontPreview'
